# perf(api): parallelize board data fetching