# What's changing?

* Replace sequential `await fetch()` calls with `Promise.all()` for parallel execution. 

Apply to all board types: regular boards, all-notes, and archive views

---

Current Sequential Pattern (Production):
```bash
  1. user        → 118ms
  2. boards      → 112ms
  3. notes       → 128ms
  4. all-notes   → 115ms
```
Total API time: 
**~600ms** just for the core data requests

Instead of this waterfall, we'd have:
```bash
  1. user        → 118ms (auth check - must be first)
  2. [boards, notes, all-notes] → max(112ms, 128ms, 115ms) = 128ms (parallel)
```
Total API time: 
**~246ms** (118ms + 128ms)

Improvement: 
**~60% faster**